// types/interfaces.ts
// properties from the API response 
export interface CryptoData {
    id: string;
    name: string;
    symbol: string;
    image: string;
  }

export interface Currency {
    id: string;
    name: string;
    symbol: string;
    image : {
        thumb: string;
        small: string;
        large: string;
    };
    current_price: number;
    market_cap: number;
    market_cap_rank: number;
    fully_diluted_valuation: number;
    total_volume: number;
    high_24h: number;
    low_24h: number;
    price_change_24h: number;
    price_change_percentage_24h: number;
    market_cap_change_24h: number;
    market_cap_change_percentage_24h: number;
    circulating_supply: number;
    total_supply: number;
    max_supply: number;
    ath: number;
    ath_change_percentage: number;
    ath_date: string;
    atl: number;
    atl_change_percentage: number;
    atl_date: string;
    roi: null;
    last_updated: string;
    description: {
        en: string;
    };
    asset_platform_id: string;
    block_time_in_minutes : number;
    categories: string[];
    additional_notices: any[];
  }